// src/features/goals/domain/models/ProgressTypes.ts
export interface DailyProgress {
  consumedCalories: number;
  remainingCalories: number;
  proteinConsumed: number;
  proteinRemaining: number;
  carbsConsumed: number;
  carbsRemaining: number;
  fatConsumed: number;
  fatRemaining: number;
  isOverCalories: boolean;
  isOverProtein: boolean;
  isOverCarbs: boolean;
  isOverFat: boolean;
}